Overview
The algorithm would use a quantum computer to generate encryption keys based on the principles of quantum mechanics. These keys could then be used to encrypt data stored in the cloud. Due to the properties of quantum mechanics, any attempt to intercept or tamper with the keys would be immediately detectable.